move lispy tests to engine.rs
