+++ stack-core/src/lib.rs
@@ -27,5 +27,4 @@ pub mod prelude {
   pub use symbol::Symbol;
 }
 
-mod tests;
 mod vec_one;
